[IMP] runbot_merge: formalize queues a bit
