+++ runbot_merge/models/crons/__init__.py
@@ -1,6 +1,9 @@
+from . import queue
 from . import cleanup_scratch_branches
 from . import issues_closer
 from . import forwardport
+from . import branch_remover
+from . import update_followup
 from . import validate
 from . import status_request
 from . import check_commits
